requirements.txt
python-telegram-bot
psycopg2-binary
matplotlib
pandas
scikit-learn
python-dotenv
xlsxwriter
numpy
schedule 